refactor: keep Rust properties outside library sources
